refactor: isolate X.509 workload identity state
